pyproject.toml
[project]
name = "Boom-Tetris"
version = "0.1.0"
description = "Boom! Tetris!"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.5",
    "pygame>=2.6.1",
    "ruamel-yaml>=0.18.11",
]

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "ruff>=0.11.6",
]

[[tool.mypy.overrides]]
module = ["yaml.*"]
follow_untyped_imports = true
